requirements.txt
# Core dependencies
python-dotenv>=0.19.0

# Speech recognition
SpeechRecognition>=3.8.1
pyaudio>=0.2.11

# AI/NLP
nltk>=3.6.0
transformers>=4.20.0
torch>=1.12.0

# Email handling
smtplib-ssl>=1.0.0
email-validator>=1.1.3

# Utilities
python-dateutil>=2.8.2
